Migrate GroovuinoML to Gradle + bump to Java 17 LTS/Groovy 4 + scripting
Use shadowJar plugin to create a fat jar with all dependencies, upgrade to latest available versions of Java and Groovy, validated behavior, added automated build and run scripts
