--- embeddeds/groovy/GroovuinoML/build.bat
@@ -1,1 +1,0 @@
-mvn clean compile assembly:single
