--- embeddeds/groovy/GroovuinoML/run.bat
@@ -1,1 +1,0 @@
-java -jar target\dsl-groovy-1.0-jar-with-dependencies.jar scripts\Switch.groovy > result.ino
