+++ embeddeds/groovy/build.bat
@@ -1,0 +1,1 @@
+gradlew.bat shadowJar
