+++ embeddeds/groovy/build.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+CWD=$(pwd)
+cd ../../kernels/jvm || exit
+mvn install
+cd "$CWD" || exit
+./gradlew shadowJar
