+++ embeddeds/groovy/scripts/Switch.groovy
