+++ embeddeds/groovy/scripts/Switch2.groovy
